refactor: Seperate packages by function for future scalability
- Split classes about Paging into distinct packages to enhance modularity and maintainability
- Prepared structure for easier feature additions in the future
