+++ src/main/java/com/ssafy/sandbox/paging/dto/PageTodoOffsetResponse.java
@@ -1,5 +1,6 @@
-package com.ssafy.sandbox.todo.dto;
+package com.ssafy.sandbox.paging.dto;
 
+import com.ssafy.sandbox.todo.dto.TodoDto;
 import lombok.Builder;
 import lombok.Getter;
